Rework of the default block content feature
Introduces a new parameter to the `flatblock` tag: 'with-default'. When
this parameter is present, the tag parses the template until the
`end_flatblock` tag is found, and everything between the tags is saved
for later rendering and used as default content for the FlatBlock
instance if it doesn't exists.

Also introduces two new settings:
`FLATBLOCKS_STRICT_DEFAULT_CHECK`: If True, when the tag founds a
Flatblock instance that matches the slug but its fields are empty,
the tag fields the instance's fields with the defaults values, if
provided.
`FLATBLOCKS_STRICT_DEFAULT_CHECK_UPDATE`: If True and the above
situation occurs then the instance is saved to the DB with the
defaults values.